refactor: refactor admin navbar structure and add NavbarItems type
